Plot China Map with Custom Projection

A ggplot2 layer to plot China's provincial map with a customizable projection.

Usage

geom_mapcn(
  data = NULL,
  mapping = aes(),
  color = "grey",
  fill = "white",
  crs = "+proj=aeqd +lat_0=35 +lon_0=105 +ellps=WGS84 +units=m +no_defs",
  size = 0.2,
  ...
)

Source

China map data provided by Tianditu, available at https://cloudcenter.tianditu.gov.cn/administrativeDivision

Arguments

data

sf object with China's map data, or NULL to load the package's default map.

color

Border color for provinces. Default is "black".

fill

Fill color for provinces. Default is "white".

crs

The coordinate reference system (CRS) for the projection. Defaults to "+proj=aeqd +lat_0=35 +lon_0=105 +ellps=WGS84 +units=m +no_defs". Users can specify other CRS strings (e.g., "+proj=merc" for Mercator) or use EPSG codes (e.g., "epsg:4326") to customize the projection.

size

Numeric. The width of the border lines for provinces. Default is 0.5.

...

Additional parameters for geom_sf.

Value

A ggplot2 layer of China's provincial boundaries.

Examples

if (FALSE) { # \dontrun{
# Default map
ggplot() +
  geom_mapcn()

# Load the China provincial map data from the package
china_path <- system.file("extdata", "China_sheng.geojson", package = "ggmapcn")
china_data <- sf::st_read(china_path, quiet = TRUE)

# Use Albers Equal Area Conic projection
ggplot() +
  geom_mapcn(data=china_data, crs = "+proj=aea +lat_1=25 +lat_2=47 +lon_0=110")+
  theme_bw() # Apply a black-and-white theme for a clean look
} # }