Modern Pollen Sampling Sites in China

This dataset contains information on 4,497 modern pollen sampling sites across China, including geographic coordinates, elevation, and sample types. It does not contain specific pollen data but serves as a valuable resource for identifying locations where modern pollen samples were collected. This information can be used to study vegetation patterns and environmental reconstructions by correlating with other data sources.

Usage

pollen

Format

An object of class tbl_df (inherits from tbl, data.frame) with 4497 rows and 21 columns.

References

Chen, H.-Y., Xu, D.-Y., Liao, M.-N., Li, K., Ni, J., Cao, X.-Y., Cheng, B., Hao, X.-D., Kong, Z.-C., Li, S.-F., Li, X.-Q., Liu, G.-X., Liu, P.-M., Liu, X.-Q., Sun, X.-J., Tang, L.-Y., Wei, H.-C., Xu, Q.-H., Yan, S., Yang, X.-D., Yang, Z.-J., Yu, G., Zhang, Y., Zhang, Z.-Y., Zhao, K.-L., Zheng, Z., Herzschuh, U. (2021). A modern pollen dataset of China. Chinese Journal of Plant Ecology, 45(7), 799–808. doi:10.17521/cjpe.2021.0024 .

Examples

data(pollen)
head(pollen)
#> # A tibble: 6 × 21
#>     No. Site             `Site name` Longitude Latitude `Elevation (m)` 样品类型
#>   <dbl> <chr>            <chr>           <dbl>    <dbl>           <dbl> <chr>   
#> 1     1 艾比湖           Aibi Lake        82.8     45               194 湖泊沉… 
#> 2     2 艾比湖 ZK00B     Aibi Lake …      82.9     44.6             195 钻孔/剖…
#> 3     3 阿拉善高原-00-02 Alashan-00…     101.      41.7             910 土壤表层
#> 4     4 阿拉善高原-00-04 Alashan-00…     101.      42.3             910 土壤表层
#> 5     5 阿拉善高原-00-05 Alashan-00…     101.      42.4             900 土壤表层
#> 6     6 阿拉善高原-00-06 Alashan-00…     101.      42.2             930 土壤表层
#> # ℹ 14 more variables: `Sample type` <chr>, 数据来源 <chr>,
#> #   `Data source` <chr>, 数据类型 <chr>, `Data type` <chr>,
#> #   `现代植被信息(野外观测/文献摘录)` <chr>,
#> #   `Modern vegetation information (Field observation / literature extract)` <chr>,
#> #   `现代植被信息(植被图提取)` <chr>,
#> #   `Modern vegetation information (the Vegetation Map of China)` <chr>,
#> #   `现代植被(样点或周边)` <chr>, …